Print error code; WIP

+++ adsApp/src/adsCom.h
@@ -83,7 +83,7 @@ typedef struct {
   char bBusy;
 } STAXISSTATUSSTRUCT;
 
-int adsReadByName(uint16_t amsPort,const char *variableAddr,ecmcOutputBufferType *outBuffer);
+long adsReadByName(uint16_t amsPort,const char *variableAddr,ecmcOutputBufferType *outBuffer);
 int adsReadByGroupOffset(uint16_t amsPort,SYMINFOSTRUCT *info, ecmcOutputBufferType *outBuffer);
 int adsWriteByName(uint16_t amsPort,const char *variableAddr,char *asciiValueToWrite,ecmcOutputBufferType *outBuffer);
 int adsWriteByGroupOffset(uint16_t amsPort,uint32_t group, uint32_t offset,uint16_t dataType,uint32_t dataSize,const char *asciiValueToWrite,ecmcOutputBufferType *asciiResponseBuffer);
